+++ src/main/kotlin/com/team1/mvp_test/domain/mvptest/service/MvpTestService.kt
@@ -152,7 +152,6 @@ class MvpTestService(
 
         if (test.recruitType == RecruitType.FIRST_COME) {
             val lock = redissonService.getLock("applyToMvpTest:$testId", 2000, 6000)
-
             val recruitCount = memberTestRepository.countByTestIdAndState(testId, MemberTestState.APPROVED)
             check(recruitCount < test.recruitNum) { MvpTestErrorMessage.TEST_ALREADY_FULL.message }
             MemberTest(
