+++ src/main/kotlin/com/team1/mvp_test/domain/step/service/StepService.kt
@@ -4,11 +4,9 @@ import com.team1.mvp_test.common.error.MvpTestErrorMessage
 import com.team1.mvp_test.common.error.StepErrorMessage
 import com.team1.mvp_test.common.exception.ModelNotFoundException
 import com.team1.mvp_test.common.exception.NoPermissionException
-import com.team1.mvp_test.domain.member.repository.MemberTestRepository
 import com.team1.mvp_test.domain.mvptest.model.MvpTestState
 import com.team1.mvp_test.domain.mvptest.repository.MvpTestRepository
 import com.team1.mvp_test.domain.report.model.ReportState
-import com.team1.mvp_test.domain.report.repository.ReportRepository
 import com.team1.mvp_test.domain.step.dto.*
 import com.team1.mvp_test.domain.step.model.Step
 import com.team1.mvp_test.domain.step.repository.StepRepository
@@ -23,8 +21,7 @@ class StepService(
     private val stepRepository: StepRepository,
     private val mvpTestRepository: MvpTestRepository,
     private val s3Service: S3Service,
-    private val memberTestRepository: MemberTestRepository,
-    private val reportRepository: ReportRepository,
+    private val memberReportService: MemberReportService
 ) {
     fun getStepList(enterpriseId: Long, testId: Long): List<StepListResponse> {
         val mvpTest = mvpTestRepository.findByIdOrNull(testId)
@@ -97,30 +94,22 @@ class StepService(
             ?: throw ModelNotFoundException("Step", stepId)
         if (enterpriseId != step.mvpTest.enterpriseId)
             throw NoPermissionException(MvpTestErrorMessage.NOT_AUTHORIZED.message)
-        //테스트 참여자, 해당 step의 리포트 한꺼번에 불러오기
-        val members = memberTestRepository.findAllAndMemberByTestId(step.mvpTest.id!!)
-        val reports = reportRepository.findAllByStepId(stepId)
-        //각각의 참여자, 리포트를 memberId로 묶어주고 반환 형태 만들기
-        val reportMap = reports.associateBy({ it.memberTest.member.id }, { it.state })
-        val reportStatusResponse = members.map { memberTest ->
-            val reportState = reportMap[memberTest.member.id] ?: ReportState.MISSING
-            ReportStatusResponse(
-                memberId = memberTest.member.id!!,
-                memberEmail = memberTest.member.email,
-                completionState = reportState
-            )
-        }
-        //참여율 백분율로 계산
-        val approvedCount = reportStatusResponse.count { it.completionState == ReportState.APPROVED }
-        val totalCount = reportStatusResponse.size
-        val completionRate = if (totalCount > 0) {
+
+        val reportStateList = memberReportService.getMappedReport(step)
+
+        return StepOverviewResponse(
+            completionRate = reportStateList.calculateProgressPercentage(),
+            reportStatusList = reportStateList
+        )
+    }
+
+    fun List<ReportStatusResponse>.calculateProgressPercentage(): Int {
+        val approvedCount = this.count { it.completionState == ReportState.APPROVED }
+        val totalCount = this.size
+        return if (totalCount > 0) {
             ((approvedCount.toDouble() / totalCount) * 100).toInt()
         } else {
             0
         }
-        return StepOverviewResponse(
-            completionRate = completionRate,
-            reportStatusList = reportStatusResponse
-        )
     }
 }
